+++ dataflow/common/__init__.py
@@ -1,4 +1,3 @@
 from . import cli
-from . import filereader
 from . import logblocks
 from . import logger